Move fraction stuff to extra class, move fit object classes to new file. Fix types in Template
